+++ pydidas/core/utils/image_utils.py
@@ -27,10 +27,15 @@ __status__ = "Production"
 __all__ = ["calculate_histogram_limits"]
 
 
+from numbers import Real
+
 import numpy as np
 
 from ..exceptions import UserConfigError
 from ..pydidas_q_settings import PydidasQsettings
+
+
+__NUM_BINS = 16384
 
 
 def calculate_histogram_limits(data: np.ndarray) -> tuple:
@@ -47,66 +52,102 @@ def calculate_histogram_limits(data: np.ndarray) -> tuple:
     tuple[float, float]
         A tuple with the low and high limits for the histogram display
     """
+    _data_min = data.min()
+    _data_max = data.max()
+
     _qsettings = PydidasQsettings()
 
-    _fraction_low = _qsettings.value("user/histogram_outlier_fraction_low", dtype=float)
-    _fraction_high = 1 - _qsettings.value(
+    _threshold_low = _qsettings.value(
+        "user/histogram_outlier_fraction_low", dtype=float
+    )
+    _threshold_high = 1 - _qsettings.value(
         "user/histogram_outlier_fraction_high", dtype=float
     )
-    if _fraction_high - _fraction_low <= 0:
+    if _threshold_high - _threshold_low <= 0:
         raise UserConfigError(
             "The selected outlier fractions are too large. No data left to display."
         )
     _cmap_limit_low = None
     _cmap_limit_high = None
     _filtered_data = data[np.isfinite(data)]
-    if _fraction_high < 1:
-        # _counts, _edges = np.histogram(_filtered_data, bins=4096)
-        # _cumcounts = np.cumsum(_counts / _filtered_data.size)
-        # _found_indices = np.where(_cumcounts <= _fraction_high)[0]
-        # _index_low = 0 if _found_indices.size == 0 else _found_indices[-1]
-        # _index_high = min(_counts.size, _index_low + 1)
-        # print(_index_low, _index_high, _cumcounts[_index_low], _cumcounts[_index_high])
-        #
-        # _counts_fine, _edges_fine = np.histogram(
-        #     _filtered_data, bins=4096, range=(_edges[_index_low], _edges[_index_high])
-        # )
-        # _cumcounts_fine = np.cumsum(_counts_fine / _filtered_data.size) + _cumcounts[_index_low]
-        # _found_indices = np.where(_cumcounts_fine <= _fraction_high)[0]
-        # _index_low_fine = 0 if _found_indices.size == 0 else _found_indices[-1]
-        # _cmap_limit_high = _edges_fine[_index_low_fine]
-        # Initial histogram to find the coarse range
-        _counts, _edges = np.histogram(_filtered_data, bins=4096)
-        _cumcounts = np.cumsum(_counts / _filtered_data.size)
-        _index_low = np.max(np.where(_cumcounts <= _fraction_high)[0], initial=0)
-
-        _counts_fine, _edges_fine = np.histogram(
-            _filtered_data,
-            bins=4096,
-            range=(_edges[_index_low], _edges[min(_index_low + 1, _counts.size)]),
+    if _threshold_high < 1:
+        _cmap_limit_high = __calc_upper_limit(_filtered_data, _threshold_high)
+    if _threshold_low > 0:
+        _cmap_limit_low = __calc_lower_limit(
+            _filtered_data, _threshold_low, _cmap_limit_high
         )
-        _cumcounts_fine = np.cumsum(_counts_fine / _filtered_data.size) + _cumcounts[_index_low]
-        _index_low_fine = np.max(np.where(_cumcounts_fine <= _fraction_high)[0], initial=0)
-        _cmap_limit_high = _edges_fine[_index_low_fine]
-    if _fraction_low > 0:
-        _range = (
-            _filtered_data.min(),
-            _cmap_limit_high if _cmap_limit_high is not None else _filtered_data.max(),
-        )
-        print(_range)
-        _counts, _edges = np.histogram(_filtered_data, bins=4096, range=_range)
-        _cumcounts = np.cumsum(_counts / _filtered_data.size)
-        print(_cumcounts)
-        _index_stop_edge = min(
-            _counts.size, np.where(_fraction_low <= _cumcounts)[0][0] + 1
-        )
-
-        _counts_fine, _edges_fine = np.histogram(
-            data, bins=32768, range=(_edges[0], _edges[_index_stop_edge])
-        )
-        _cumcounts_fine = np.cumsum(_counts_fine / _filtered_data.size)
-        _index_stop_edge_fine = min(
-            _counts_fine.size, np.where(_fraction_low <= _cumcounts_fine)[0][0] + 1
-        )
-        _cmap_limit_low = _edges_fine[_index_stop_edge_fine]
+    if isinstance(_cmap_limit_low, Real) and _cmap_limit_low > _data_max:
+        _cmap_limit_low = _data_min - 1e-3 * (_data_max - _data_min)
+    if isinstance(_cmap_limit_high, Real) and _cmap_limit_high < _data_min:
+        _cmap_limit_high = _data_max + 1e-3 * (_data_max - _data_min)
     return _cmap_limit_low, _cmap_limit_high
+
+
+def __calc_upper_limit(data: np.ndarray, threshold_high: float) -> float:
+    """
+    Calculate the upper limit for the histogram display.
+
+    Parameters
+    ----------
+    data : np.ndarray
+        The image data to calculate the histogram limits for.
+    threshold_high : float
+        The relative threshold of high data values data to be ignored.
+
+    Returns
+    -------
+    float
+        The upper limit for the histogram display.
+    """
+    _counts, _edges = np.histogram(data, bins=__NUM_BINS)
+    _cumcounts = np.cumsum(_counts / data.size)
+    _bin_index_low = np.max(np.where(_cumcounts <= threshold_high)[0], initial=0)
+    try:
+        # new range must be range of bin following low bin:
+        _new_range = (_edges[_bin_index_low + 1], _edges[_bin_index_low + 2])
+    except IndexError:
+        _new_range = (_edges[_bin_index_low], _edges[_bin_index_low + 1])
+    _counts_fine, _edges_fine = np.histogram(data, bins=__NUM_BINS, range=_new_range)
+    _cumcounts_fine = np.cumsum(_counts_fine / data.size) + _cumcounts[_bin_index_low]
+    _index_new = np.max(np.where(_cumcounts_fine <= threshold_high)[0], initial=0)
+    return _edges_fine[_index_new + 1]
+
+
+def __calc_lower_limit(
+    data: np.ndarray, threshold_low: float, cmap_high_lim: float
+) -> float:
+    """
+    Calculate the lower limit for the histogram display.
+
+    Parameters
+    ----------
+    data : np.ndarray
+        The image data to calculate the histogram limits for.
+    threshold_low : float
+        The relative threshold of low data values data to be ignored.
+    cmap_high_lim : float
+        The upper limit for the histogram display.
+
+    Returns
+    -------
+    float
+        The lower limit for the histogram display.
+    """
+    _range = (
+        data.min(),
+        cmap_high_lim if cmap_high_lim is not None else data.max(),
+    )
+    if _range[0] >= _range[1]:
+        _range = (_range[0] - 1e-4, _range[0] + 1e-4)
+    _counts, _edges = np.histogram(data, bins=__NUM_BINS, range=_range)
+    _cumcounts = np.cumsum(_counts / data.size)
+    _bin_index_fine = np.max(np.where(_cumcounts <= threshold_low)[0], initial=0)
+    _new_range = (_edges[_bin_index_fine], _edges[_bin_index_fine + 1])
+    _counts_fine, _edges_fine = np.histogram(
+        data,
+        bins=__NUM_BINS,
+        range=(_edges[_bin_index_fine], _edges[_bin_index_fine + 1]),
+    )
+    _cumcounts_fine = np.cumsum(_counts_fine / data.size) + _cumcounts[_bin_index_fine]
+    _index_new = np.min(np.where(_cumcounts_fine >= threshold_low)[0], initial=0)
+    return _edges_fine[_index_new]