+++ tests/core/utils/test_image_utils.py
@@ -32,7 +32,6 @@ from pydidas.core import PydidasQsettings, UserConfigError
 from pydidas.core.utils import (
     calculate_histogram_limits,
 )
-from pydidas.data_io import import_data
 
 
 class TestImageUtils(unittest.TestCase):
@@ -72,10 +71,9 @@ class TestImageUtils(unittest.TestCase):
         raw_data = np.arange(1000**2)
         self.qsettings.set_value("user/histogram_outlier_fraction_low", 0)
         self.qsettings.set_value("user/histogram_outlier_fraction_high", 1 - 1e-6)
-        data = raw_data
-        low, high = calculate_histogram_limits(data)
+        low, high = calculate_histogram_limits(raw_data)
         self.assertIsNone(low)
-        self.assertAlmostEqual(high, 0)
+        self.assertTrue(high < 500)
 
     def test_calculate_histogram_limits__constant_arr_values(self):
         raw_data = np.ones(100**2)
@@ -83,9 +81,8 @@ class TestImageUtils(unittest.TestCase):
         self.qsettings.set_value("user/histogram_outlier_fraction_high", 0.05)
         data = raw_data
         low, high = calculate_histogram_limits(data)
-        print("flat data: ", low, high)
-        # self.assertTrue(abs(low - (500 + _offset)) < 25)
-        # self.assertTrue(abs(high - (9500 + _offset)) < 25)
+        self.assertTrue(abs(low - 1) < 1e-4)
+        self.assertTrue(abs(high - 1) < 1e-4)
 
     def test_calculate_histogram_limits__no_upper_limit(self):
         raw_data = np.arange(100**2)
@@ -114,23 +111,13 @@ class TestImageUtils(unittest.TestCase):
         raw_data[0] = -1e6
         self.qsettings.set_value("user/histogram_outlier_fraction_low", 0.05)
         self.qsettings.set_value("user/histogram_outlier_fraction_high", 0.05)
-        for _offset in [0 , -500, -5e4, 12000, 120000]:
+        for _offset in [0, -500, -5e4, 12000, 120000]:
             with self.subTest(offset=_offset):
                 _tolerance = 400 * (1 + abs(_offset) / 1e5)
                 data = raw_data + _offset
                 low, high = calculate_histogram_limits(data)
                 self.assertTrue(abs(low - (500 + _offset)) < _tolerance)
                 self.assertTrue(abs(high - (9500 + _offset)) < _tolerance)
-
-    def test_calculate_histogram_limits__real_data(self):
-        raw_data = import_data(r"E:\_data\David_Canelo\T6_Malte\T6_Centre-00141.tif")
-        self.qsettings.set_value("user/histogram_outlier_fraction_low", 0.05)
-        self.qsettings.set_value("user/histogram_outlier_fraction_high", 0.05)
-        _tolerance = 400
-        data = raw_data
-        low, high = calculate_histogram_limits(data)
-        # self.assertTrue(abs(low - (500 + _offset)) < _tolerance)
-        # self.assertTrue(abs(high - (9500 + _offset)) < _tolerance)
 
 
 if __name__ == "__main__":